Extract columns and rejections gen actions as top-level types
