+++ tools/model-compiler/src/main/java/io/spine/tools/gradle/compiler/RejectionGenPlugin.java
@@ -19,30 +19,16 @@
  */
 package io.spine.tools.gradle.compiler;
 
-import com.google.common.collect.ImmutableSet;
-import io.spine.code.gen.Indent;
-import io.spine.code.java.PackageName;
-import io.spine.code.java.SimpleClassName;
 import io.spine.code.proto.FileSet;
-import io.spine.code.proto.RejectionsFile;
-import io.spine.code.proto.SourceProtoBelongsToModule;
-import io.spine.tools.compiler.gen.GeneratedTypeSpec;
-import io.spine.tools.compiler.gen.rejection.RejectionSpec;
-import io.spine.tools.gradle.CodeGenerationAction;
 import io.spine.tools.gradle.GradleTask;
 import io.spine.tools.gradle.ProtoPlugin;
-import io.spine.type.RejectionType;
 import org.gradle.api.Action;
 import org.gradle.api.Project;
 import org.gradle.api.Task;
 
 import java.io.File;
-import java.util.List;
 import java.util.function.Supplier;
 
-import static com.google.common.collect.ImmutableSet.toImmutableSet;
-import static com.google.common.flogger.LazyArgs.lazy;
-import static io.spine.code.proto.RejectionsFile.findAll;
 import static io.spine.tools.gradle.JavaTaskName.compileJava;
 import static io.spine.tools.gradle.JavaTaskName.compileTestJava;
 import static io.spine.tools.gradle.ModelCompilerTaskName.generateRejections;
@@ -113,7 +99,7 @@ public class RejectionGenPlugin extends ProtoPlugin {
                                              Supplier<FileSet> files,
                                              Supplier<String> targetDirPath,
                                              Supplier<String> protoSrcDir) {
-        return new GenAction(project, files, targetDirPath, protoSrcDir);
+        return new RejectionGenAction(project, files, targetDirPath, protoSrcDir);
     }
 
     @Override
@@ -124,72 +110,5 @@ public class RejectionGenPlugin extends ProtoPlugin {
     @Override
     protected Supplier<File> testDescriptorFile(Project project) {
         return () -> getTestDescriptorSet(project);
-    }
-
-    /**
-     * Generates source code of rejections.
-     */
-    private static class GenAction extends CodeGenerationAction {
-
-        private GenAction(Project project,
-                          Supplier<FileSet> files,
-                          Supplier<String> targetDirPath,
-                          Supplier<String> protoSrcDirPath) {
-            super(project, files, targetDirPath, protoSrcDirPath);
-        }
-
-        @Override
-        public void execute(Task task) {
-            ImmutableSet<RejectionsFile> rejectionFiles = findModuleRejections(protoFiles().get());
-            _debug().log("Processing the file descriptors for the rejections `%s`.",
-                         rejectionFiles);
-            for (RejectionsFile source : rejectionFiles) {
-                // We are sure that this is a rejections file because we got them filtered.
-                generateRejections(source);
-            }
-        }
-
-        /**
-         * Obtains all rejection files belonging to the currently processed module.
-         */
-        private ImmutableSet<RejectionsFile> findModuleRejections(FileSet allFiles) {
-            ImmutableSet<RejectionsFile> allRejections = findAll(allFiles);
-            ImmutableSet<RejectionsFile> moduleRejections = allRejections
-                    .stream()
-                    .filter(new SourceProtoBelongsToModule(protoSrcDir()))
-                    .collect(toImmutableSet());
-            return moduleRejections;
-        }
-
-        private void generateRejections(RejectionsFile source) {
-            List<RejectionType> rejections = source.rejectionDeclarations();
-            if (rejections.isEmpty()) {
-                return;
-            }
-
-            logGeneratingForFile(source);
-            for (RejectionType rejectionType : rejections) {
-                // The name of the generated `ThrowableMessage` will be the same
-                // as for the Protobuf message.
-                _debug().log("Processing rejection `%s`.", rejectionType.simpleJavaClassName());
-                GeneratedTypeSpec spec = new RejectionSpec(rejectionType);
-                spec.writeToFile(targetDir().toPath(), indent());
-            }
-        }
-
-        private void logGeneratingForFile(RejectionsFile source) {
-            _debug().log(
-                    "Generating rejections from the file: `%s` " +
-                            "`javaPackage`: `%s`, `javaOuterClassName`: `%s`.",
-                    source.path(),
-                    lazy(() -> PackageName.resolve(source.descriptor().toProto())),
-                    lazy(() -> SimpleClassName.outerOf(source.descriptor()))
-            );
-        }
-
-        @Override
-        protected Indent getIndent(Project project) {
-            return Extension.getIndent(project);
-        }
     }
 }
